CI: Updates
- option not to download CUDA [Win]
- correct asset uploading
- nightly CI: use different trigger - temporarily "on push" (to
  ci-nightly)
- correctly add DLLs (using "|" as a delimiter to cope with spaces in
  names)
- extract to C:\
- remove correct test artefact [Win]
- build AJA wrapper only if we have SDK [Win]
- disable NDI by default

+++ .github/scripts/Windows/prepare.ps1
@@ -9,10 +9,12 @@ choco install --no-progress msys2 --params "/NoUpdate /InstallDir:C:\msys64"
 echo "::set-env name=MSYS2_PATH_TYPE::inherit" # MSYS2 inherits PATH from Windows
 
 # Install CUDA
-Invoke-WebRequest https://developer.download.nvidia.com/compute/cuda/10.2/Prod/local_installers/cuda_10.2.89_441.22_win10.exe -OutFile cuda_inst.exe
-Start-Process -FilePath "cuda_inst.exe" -ArgumentList "-s nvcc_10.2" -Wait -NoNewWindow
-Remove-Item cuda_inst.exe
-echo "::add-path::C:\Program Files\NVIDIA GPU Computing Toolkit\CUDA\v10.2\bin"
+if (!${env:no_cuda}) {
+  Invoke-WebRequest https://developer.download.nvidia.com/compute/cuda/10.2/Prod/local_installers/cuda_10.2.89_441.22_win10.exe -OutFile cuda_inst.exe
+  Start-Process -FilePath "cuda_inst.exe" -ArgumentList "-s nvcc_10.2" -Wait -NoNewWindow
+  Remove-Item cuda_inst.exe
+  echo "::add-path::C:\Program Files\NVIDIA GPU Computing Toolkit\CUDA\v10.2\bin"
+}
 
 # Build AJA
 if (${env:sdk_pass}) {
@@ -21,7 +23,7 @@ if (${env:sdk_pass}) {
   $basicAuthValue = "Basic $encodedCreds"
   $Headers = @{Authorization = $basicAuthValue}
   Invoke-WebRequest -Headers $Headers https://frakira.fi.muni.cz/~xpulec/sdks/ntv2sdkwin.zip -OutFile aja.zip
-  Expand-Archive -LiteralPath 'aja.zip' -DestinationPath 'C:'
+  Expand-Archive -LiteralPath 'aja.zip' -DestinationPath 'C:\'
   Remove-Item aja.zip
   mv c:\ntv2sdk* c:\AJA
   cd c:\AJA
@@ -35,7 +37,7 @@ if (${env:sdk_pass}) {
   $basicAuthValue = "Basic $encodedCreds"
   $Headers = @{Authorization = $basicAuthValue}
   Invoke-WebRequest -Headers $Headers https://frakira.fi.muni.cz/~xpulec/sdks/ximea.zip -OutFile ximea.zip
-  Expand-Archive -LiteralPath 'ximea.zip' -DestinationPath 'C:'
+  Expand-Archive -LiteralPath 'ximea.zip' -DestinationPath 'C:\'
   Remove-Item ximea.zip
 }
